Update and rename action.yml to hello-to-news-contributions.yml
Here's the configuration for the `.github/workflows/hello-to-news-contributions.yml` file you requested:

This workflow will automatically run whenever a pull request or issue is opened in the repository. The `actions/first-interaction` action is used to check whether the user opening the pull request or issue is a first-time contributor. If they are, the specified message will be posted as a comment on their pull request or issue. The `${{ secrets.GITHUB_TOKEN }}` expression is used to authenticate the action with the repository's secret token.